pyproject.toml
[tool.poetry]
name = "prescription-automation"
version = "0.1.0"
description = ""
authors = ["Robin Lightfoot"]
readme = "README.md"
packages = [{include = "prescription_automation"}]

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.69.0"
constructs = ">=10.0.0,<11.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
